Symforce generated geometry types for C++ and Python
This does not include any generation code, just the outputs. The
idea is to allow actually using the types to provide feedback and
begin integration in parallel with code reviewing the generation
side. Eventually this code may not be checked in but generated by
the build. This change will be a no-op to users.

See symforce_geo_test.cc for usage examples. Basic use:
#include <geo/rot3.h> --> geo::Rot3d() or geo::Rot3f()
import geo --> geo.Rot3()

Feature-Branch: feature/symforce_geo_types
Base-Branch: master