Symforce fixed size Matrix strategy
This properly fits the ops concepts that expect @classmethods to
know about their dimension, and roughly parallels Eigen in C++. It
also means you can do geo.M32.zero() instead of geo.M32().zero().

If this works, we can go into the ops and remove all the (Element or type)
crap which is meant to handle this use case where you need the instance.

We also should take a pass through notebooks and get rid of the extra parens.

The general way this is accomplished is by overriding __new__ to return
fixed size child classes instead of Matrix itself, which then statically
set MATRIX_DIMS. The risk is that some methods in sympy or symengine will
change the shape of the matrix object without constructing a new one. I have
a guess that functions like row_join, col_join, row_del might be in that
category. It doesn't seem terrible to handle these as we hit them though.

[Testing] Added detailed test cases for construction methods.

Feature-Branch: symforce-matrix-constructor-helpers
Base-Branch: master