Run clang-format on SymForce sources and generated code
This runs clang-format as part of `make format` and `make check_format` in SymForce, as well as on any C++ code generated by SymForce, including by other packages.  Currently it uses SymForce's .clang-format to do this, so there will be an issue if this diverges from Skydio's .clang-format.  Also does not run black on generated python code; the best way to do this would be to call black from python, but black is Python3 only so this requires running SymForce with Python3.

[Testing] Existing unit tests pass

Feature-Branch: aaron/symforce-autoformat-cpp
Base-Branch: master