Fix SymForce `make test`
`make test` passes again.  This includes running (most) tests on both the symengine and sympy backends when running `make test`.

This also adds the test_symengine and test_sympy make targets to run tests on only one backend, as well as the --run_slow_tests flag that can be passed to a test to run everything when executing with the sympy backend (it takes about 20 minutes right now to run all tests this way)

[Testing] Existing tests

Feature-Branch: aaron/symforce-fix-make-test
Base-Branch: master