[SymForce] Bundle adjustment example
There are definitely some gross things in here, especially in the fixed-size problem. I'm proposing that we merge as-is (except for style things, things that are fixable without core SymForce changes, etc) so we have a reference for what this currently looks like - it'll also mean this doesn't keep getting dragged along on my workstation until we get the example nice and clean.