add constant input to measurement function of SymForce EKF
This change adds a values object input to the measurement function
which acts as constants not part of the state. This allows for more
expressive measurement functions. It is functionally similar to the
control inputs of the process function.

The use-case in mind is the gimbal which has measurements on the
acceleration measured by the IMUs. Specifically, the state is the
rotation between the vehicle and camera. The vehicle and camera
both have accel measurements which when rotated should match.
Neither acceleration is part of the state, thus one is considered
a constant of the measurement function.

[Testing] not tested

Feature-Branch: ekf_meas_const
Base-Branch: master