[SymForce] Add numba integration
Adds the option to apply the `@numba.njit` decorator to generated python functions (along with a mechanism for passing language-specific arguments through to the templates).  This is off by default, because it'll be slower for small functions or functions that are only called a couple times.  Numba also isn't added to requirements.txt and the unit test will be skipped if it's not installed, because it seems like it shouldn't be a required dependency.
    
[Testing] Added unit test