Bradley/rot3 from rotation matrix rotate 180 test
Adding a test case to catch the error of `Rot3.from_rotation_matrix` that occurs when a rotation matrix corresponds to a 180 degree rotation about an axis whose components are not all of the same sign. The error manifests in the form of the axis components all being set to be positive (even when some should be negative).