[SymForce] Rot3.from_rotation_matrix fix
This request adds additional tests to check that `Rot3.from_rotation_matrix` is numerically stable for rotations (very) near 180 degrees. Also, changes `Rot3.from_rotation_matrix` to have correct answers for rotations of exactly 180 degrees and near 180 degrees.

Things that stick out about this to me:
-`Rot3.from_rotation_matrix` is now very long. Both because a lot of comments were needed to explain what was going on and why, but also because the implementation itself is kind of long. This is mitigated by the fact core steps were separated out into sub-functions when appropriate
-In `M33_pow` I had to type cast a 3x3 Matrix to a `Matrix33` because
the product of two `M33` matrices is a `Matrix`, and not a `M33` for
some reason, and I didn't know of a better way to handle this