Adding a test for Values which is expected to fail
Despite what the docstring suggests, the keys returned by `Values.items_recursive` are not valid keys for the object.

For example, if
`v = Values(a_list=[1, 2, 3, 4])`
`key, value = v.items_recursive()[0]`
then
`v[key]`
will return an error as `key` equals `"a_list[0]"`, which is not a key of `v`.

Assuming I'm not misunderstanding the purpose of this method, I think this is bad and should be changed.

We can either change it by making `"a_list[0]"` into a valid key, or by not flattening lists (for example, a `geo.Vector3` is not currently flattened I don't think).