[SymForce] Enable Values.from_storage_index to work on new CameraCals automatically
As things stand, `Values.from_storage_index` handles subclasses of `CameraCal` by looking up a manually made list of the names of `CameraCal` subclasses. It looks like this list was not kept up to date, as it didn't have `SphericalCameraCal`.

I primarily wanted to just add a failing test, but since it was easy to fix, I figured I'd ought to at least patch it up so that this won't be a problem in the short term.

I don't like the series of `elif`'s in `Values.from_storage_index` though, as I feel like this is what `StorageOps.from_storage` is for. But, when I looked into that, I discovered that it has several issues (don't try to store tuples for example). I plan to fix that up, but it might take longer (I'll probably need to consult with people for it too, and it's the weekend now so it will take at least that long probably).