[SymForce] lists, tuples, and ndarrays should implement LieGroupOps
When their components are all lie groups (or at least floats), you should expect the implementations for `list`s, `tuple`s, and `ndarray`s to be proper implementations of LieGroupOps. But, with the exception of `list`s, a lot of things that you would expect to work simply don't. For example,
`LieGroupOps.from_storage((0, 0), LieGroupOps.to_storage((1, 2)))`
raises
`AttributeError: 'tuple' object has no attribute 'append'`
because `SequenceStorageOps.from_storage` assumes the type has an append method. This would be understandable if a lot of types with an append method had their implementations backed by `SequenceStorageOps`, but, as far as I can tell, the only three types that use it are `list`, `tuple`, and `ndarray` (which doesn't have an append attribute either I don't think).

There's also the uncomfortable fact that while `StorageOps.to_storage` can take a type as its first argument, `SequenceStorageOps.to_storage` can't. It's understandable that we might need some info about how long the list is and its contents, but the length of the list could be included in its storage vector and the contents could be handled with something like `T.List[str]` if we have a list of strings, as we can access the element type with `T.List[str].__args__` or `T.Tuple[str].__args__` (I don't know how to get the element type from an `np.ndarray`, but perhaps we can just assume it's a `T.Scalar` or something).

In a way, this pull request is more meant to spark discussion and see if we should try to make these things more reliable.

I just got interested in it because the lack of consistent implementations of `to_storage` and `from_storage` seems to be the root cause (as far as I can tell) for some of the unpleasant aspects of `Values.index` and `Values.from_storage_index`, which is what I was working on when I got sucked into this rabbit hole.

Assert LieGroupOps for list, tuple, ndarray