[SymForce] Enable Values.from_storage_index to handle Values with DualQuaternions
A lot like a similar pull request for CameraCals, but more limited in scope.

Right now `Values.from_storage_index` will raise a `NotImplementedError` if used on the outputs of `Values.to_storage` and `Values.index` of a Values containing a `DualQuaternion`.

This pull request just adds a test for this and puts on a band aid to fix it.

Unlike the `CameraCals` version linked above, there isn't an obvious (to me at least) clean way to support any `Storage` class in `geo`. However, in conversation with Aaron, since `geo` (unlike `cam`) isn't changing that much, we thought it was okay to stick with this brittle implementation, at least for now.