[SymForce] Refactor IndexEntry
Currently, the output of `Values.index` is used in quite a few places, particularly those related to type information in codegen (`codegen_util.py`, `types_package_codgen.py`, `types.lcm.jinja`, and `lcm_templates/util.jinja`).

Since currently we just return a bare list to represent an index entry, there is very little room to modify the form of index without impacting these other files. To address this, we introduce an additional layer of abstraction by creating an `IndexEntry` class to represent this structure. This way, for instance, we can add an additional field (like `storage_dim`) without breaking existing code (which up to this point has depended on an `IndexEntry` having only 4 elements in a particular order).

The structure of `IndexEntry` is meant to mirror (more or less) that of `index_entry_t` in `symforce.lcm`.

Introducing this class has the additional benefit of making the code that uses it a bit clearer, as, rather than having to read through the implementation of `Values.index` to understand the structure of the data, now one only has to look at `IndexEntry`. Additionally, one no longer needs to know the order of elements of an `IndexEntry` in order to use and understand them.

One major thing that this pull request does _not_ cover is eliminating the need to look at the implementation of `Values.index` in order to learn what the corresponding string in `IndexEntry.datatype` is. For instance, the `dataclass` string of a `tuple` is `"List"`, of a `numpy.ndarray` is `"np.ndarray"`, of a `float` is `"Scalar"`, of a `geo.V3` is `"Matrix"`, and of a `geo.Rot3`is `"Rot3"` (the point being is that it's kind of hard to guess what it would be without already knowing the answer). This has the additional problem that one can not change the string `datatype` of a class without also having to change everywhere that string is used (this was the motivating issue, as I attempted to change the dataclass of a `tuple` to `"tuple"` instead of `"List"').

However, this class make it easier to remove the use of these strings outside of `Values` by providing a place to add methods like `is_matrix_type()` or `is_list_type()` (which is the type of information that the above listed files use) while leaving the implementer of `Values` free to change the actual strings used to represent the datatypes needed to unpack the data from storage in places like `Values.from_storage_index`. (We might also want a function which takes a type and returns the associated `datatype` string).

**A possible change**
Currently, I have `IndexEntry` subclassing `T.NamedTuple`. This is convenient, but the thing that concerns me is that `T.NamedTuple`s can be unpacked. While I originally liked this because it made it easier to swap in `IndexEntry` initially (because the existing code used a lot of unpacking), I suspect this could be bad going forward, as it allows users of `IndexEntry` to write code which depends on the order of elements in `IndexEntry`. Unpacking didn't seem to provide much dramatic benefit in the places where I saw it used while working on this, but it makes it harder to add new fields or change their order (for whatever reason). That's half of what gave me difficulty while making my changes, so I'm hesitant to allow it to happen again.

I could see annotating with `dataclasses.dataclass` instead, as that way unpacking wouldn't be allowed.