[SymForce] Treat np.ndarrays like lists in Values.items_recursive
Currently:
`Values(a=np.array([geo.Rot3.identity(), geo.Rot3.identity()])).scalar_keys_recursive()`
returns `["a[0]", "a[1]", "a[2]", "a[3]", "a[4]", "a[5]", "a[6]", "a[7]"]`
instead of `["a[0][0]", "a[0][1]", "a[0][2]", "a[0][3]", "a[1][0]", "a[1][1]", "a[1][2]" ,"a[1][3]"]`
(which is what we'd get if we used a `list` instead of an `np.ndarray`)

`Values(a=np.array([1])).scalar_keys_recursive()`
returns `["a"]`
instead of `["a[0]"]`
(which is what we'd get if we used a `list` instead of an `np.ndarray`)

`Values(a=np.array([1, 2])).items_recursive()`
returns `[("a", np.array([1, 2]))]`
instead of `[("a[0]", 1), ("a[1]", 2)]`
(which is what we'd get if we used a `list` instead of `np.ndarray`)

This pull request adds tests to ensure that when we have a `np.ndarray` in these 3 scenarios we get the `list` output, and makes a small change to get those new tests passing.

I'm not sure if this is a pull request we want to merge, as it depends entirely on what sort of behavior we'd like to see. One could make the argument that it'd be more consistent if `np.ndarray` acted like `list` in these scenarios (for example, the fact that these were treated differently surprised me when I came across it).

I'd also be open to arguments for keeping this the way they are, or only making some of the changes.