[Symforce] Add datatype recovery method to index entry
**Overview of the direction I'm trying to go**

As in the title, this is to add a `datatype` method which returns the datatype of an `IndexEntry` as a type (and not a string or anything else).

This is ultimately motivated by the realization that a lot of the pain I've been experiencing trying to straighten out the behavior of `IndexEntry` has to do with the fact that the `_datatype` (meaning the string that is used in the type codegen files) was used for two purposes 1) Reconstructing a `Values` from storage, and 2) generating the lcm types in codegen (henceforth, when I say "codegen files", I mean `codegen_util.py`, `types_package_codgen.py`, `types.lcm.jinja`, and `lcm_templates/util.jinja`)

This makes it hard to improve their use for one need without breaking their usage for the other. So, the goal is to split them apart.

Right now, most of the information stored in `_datatype` is not actually needed in `from_storage_index` (that is, if we were to have the actual datatype available). It is, however, needed in the codgen files, and `IndexEntry` and `Values` are making important decisions about what entries' `_datatype` is without knowing a thing about what sort of information the codegen files want from it.

This is awkward, and in a later pull request (that I plan to make relative to this one), I want to make all the codegen files depend only on the type returned by `datatype()` (the method I'm adding). If categorization into an enum or string would still be useful, that categorization can be done in a function defined in the files where it's actually needed (and not the distantly related values.py).

Once the codegen files are no longer dependent on `_datatype`, I'll remove references to `_datatype` from `values.py`, using only `datatype()` and the other fields of an `IndexEntry`.

This is all to reduce coupling between `Values` and the codegen files so that it's easier to fix the sorts of errors that Aaron and I have discovered while on this journey and to make it easier to understand what this code is doing (you no longer will need to understand a whole other set of files in order to understand why strange things are going on in `values.py`).

**Changes**
- Define and test `getattr_recursive` in `python_util`. This function is needed in `datatype()` so that it can handle nested classes.
- Adding tests of `Values.__eq__` and `Values.from_storage_index`. These are cannibalized from another PR. What is needed to get these tests passing is to compare indices in equality, to add differentiating type information in indices (particularly between lists and tuples), and to recover the correct type when we pull a values from storage (if we do the first two things, when we recover a `Values` from storage, it will no longer be "\_\_equal\_\_" to the value put into storage). I included these tests in this pull request because this pull request is already doing the hard part (distinguishing between different types), and because if one is not careful, if one does compare indices as well, one will break other tests (that ought to be broken) if one distinguishes between different types of scalars. So having these tests and changes now will help ensure that my changes for `datatype()` don't distinguish between different scalar types.
- Compared indices as well in `Values.__eq__`. As mentioned above, this came from a cannibalized pull request. To understand what happens if you don't compare indices, I'd recommend looking at the added tests of `__eq__`, as many `Values` which shouldn't be equal by a reasonable definition are considered equal.
- Renaming the `IndexEntry` `datatype` field to `_datatype` field (so the method can have name `datatype()` without collision).
- Changed the `__init__` method for `IndexEntry`. It used to just take all the fields of an `IndexEntry` as arguments. However, I wanted the `__module__` and `__qualname__` of the type being indexed as fields. Making the user of `IndexEntry` provide these is a bit uncomfortable as they aren't the most commonly used attributes, and if they mess it up, the breakage (in `datatype()`) could occur much later in the program execution, making it hard to debug. Since these fields aren't really needed, I thought it better to offload that work from the user to the constructor. The core difference of this new constructor is it also takes the type of the value being stored.
- Two static methods, `IndexEntry.type_encoder` and `IndexEntry.type_decoder`. `type_encoder` takes the type to be stored and encodes it as two strings, one for the module name and one for the qualified name of the type. `type_decoder` looks up the module name is the `sys.module` dictionary of loaded modules, then checks if it has a type whose keys is the type name. I don't just use `__name__` instead of `__qualname__` because the `__name__` of a nested class doesn't include the parent class (if we have a class `Outer` with nested class `Inner`, `Inner.__name__` is `"Inner"`, while `Inner.__qualname__` is `"Outer.Inner"`). The parent's name is needed to actually locate the nested class in the module. Consequently, in `type_decoder`, we must use `gettattr_recursive` because `sys.module[Inner.__module__]` does not have an attribute called `"Inner"` or `"Outer.Inner"`, but does have an attribute called `"Outer"`, and `sys.module[Inner.__module__]["Outer"]` has an attribute called `"Inner"`. So we split `Inner.__qualname__` on `"."` and recursively call `getattr` with those attribute names.
- A second note on the previous point: I originally had the implementation of `type_encoder` inlined in `__init__`, and `type_decoder` inlined in `datatype()`, but moved them into separate functions to make testing easier. The original reasons I wanted them separate ended up dissipating, but I thought having them as separate coupled functions was a lot easier to understand than having them jumbled up with the other fields, so I kept it as it.
- Adding tests of all my new `IndexEntry` methods (particularly as they relate to recovering the type). The only thing I am kind of uncomfortable about is how similar `test_encoder_and_decoder` and `test_datatype` are to each other. The only difference is that `test_datatype` doesn't use `object` as a type, because `IndexEntry`s should really only contain `StorageOps` registered types (the `storage_dim` field doesn't make sense otherwise). `test_datatype` is needed to make sure `__init__` and `datatype()` call the encoders correctly, and I thought it better to have straightforward tests of the encoders rather than just testing them indirectly.
- I also changed `index_entry_helper` and the `"Scalar"` case in `Values.get_index_from_items` to use `T.Scalar` instead of the actual scalar type. This is because Aaron convinced me that `Values` (and by extension `IndexEntry`) shouldn't distinguish between different scalar types. We only distinguish between different types explicitly so we can recover a `Values` from storage; but this isn't needed for scalars, and distinguishing between different scalar types making up a `Values` would be like saying two `geo.Rot3`s are not equal, even if they have the same numerical values, but happen to use different types to represent those values (say, `float` vs `np.float64`).

One change I didn't make was making `storage_dims` automatically calculated inside `IndexEntry.__init__` like `_module` and `_qualname`. This is because one can't infer the storage dimension of a list, for example, just from knowing that it is a list. You need an actual instance of the list so you can count the elements and see what type they are.