[SymForce] Small cleanup: remove redundant checks
Old version would continue if `list(entry.item_index.values())[0]`
was not a `list` or `values` or if
`get_subvalues_from_list_index(entry.item_index)` returned `None`.

But `get_subvalues_from_list_index(entry.item_index)` returns None
if `list(entry.item_index.values()][0]` is not a `list` or `Values`.
Hence, there is no need to check if `list(entry.item_index.values())[0]`
is a list or Values in this section of the code as it's already checked in
`get_subvalues_from_list_index`.

If this looks familiar it's because it's the recommendation Aaron made. I'm cannibalizing that pull request because I don't think the way I did things in that pull request is very good.