[SymForce] Allow indices in Values' keys
A while ago we added the test `test_items_recursive` because we wanted to allow to access indexes elements of a value in a `Values`. For example, if
`v = Values(a=[1, 2, 3])`
then we should be able to
`assert v["a[1]"] == 2`

This pull request just makes that happen.

In order to change this behavior, we needed to effectively change the implementations of `__getitem__`, `__setitem__`, `__delitem__` (in order to be consistent), and `__contains__`.

Essentially, whereas before we would grab the lowest level `Values` containing the item we were interested in, then use the most local name to access it, this change instead also indexes into the element (if it has indices) to get the indexed sub-elements.

To support this, I added the utility function `base_and_indices` which strips the indices from a key and returns what remains as the "base", and returns the indices as a list of integers.

I also modified `_get_subvalues_and_key` to also return the indices list of the most local name (and I thus renamed it to `_get_subvalues_key_and_indices`. This is the helper method that actually makes use of `base_and_indices`.

Other than that, more or less, I just added tests to ensure the modified code works as it's supposed to.