[SymForce] Remove optional index argument from from_storage
This PR is to remove the optional `index` argument from `Values.from_storage`. The reason is because the optional argument is not used anywhere, because the doc-string doesn't say anything about copying the structure of `index` (instead, it says to copy that of `self`), and if one did copy the structure of `index` instead of self, the non-existent role of `self` only creates confusion (as why would one not simply use `from_storage_index` instead?).

Incidentally, this PR removes a branch which previously had no code coverage, thereby increasing our code coverage percentage in this file.