[SymForce] Codegen python classmethods and static methods as C++ static methods
Currently it's not possible to codegen python class methods or static methods. Also, there isn't much control over default arguments.

What this PR does is enable python class methods and static methods to be codegened into C++. It also makes it so that keyword only arguments in python functions are translated as C++ methods without the parameter, with the default value always being used. (Later, it would be nice to make it so that keyword or positional arguments with a default value are codegened as default values in C++).

The motivation for this was to allow us to codegen `random_from_uniform_samples`, which is both a class method and has a keyword argument that we do not want to appear in C++.

Two changes were needed to get this working:
1) Replacing `inspect.getfullargspec` with `inspect.signature`, as the former doesn't give us much information about keyword arguments and returns `cls` as an argument even if it had already been bound, while the later omits bound arguments (which is useful as nothing needs to be done with them, but if they're around we have to know to ignore them) and tells us what type of parameter an argument is).
2) Define the macros `method_declaration` and `method_declaration_custom_namespace` to handle the decisions on whether to add `static` to the front or `const` to the back of a method declaration. This is because this logic would need to be applied anywhere we'd like to generate `static` method declarations, so it's easier to just have it in one place.