[SymForce] Define C++ ops template specifications directly rather than by inheriting from base classes
The only files that were modified directly were the `jinja` files.

Previously, we would implement the template specialization of (for example) `LieGroupOps<Rot3<Scalar>>` by defining a class `rot3::LieGroupOps<Scalar>`, then defining the template specialization `LieGroupOps<Rot3<Scalar>>` to inherit from `rot3::LieGroupOps<Scalar>`.

I am told it was done this way to get the code to compile. However, whatever seemed to be preventing compilation before seems to no longer be a problem.

All this pull request does is get rid of `rot3::LieGroupOps<Scalar>` and place it's specialization definition body in that of `LieGroupOps<Rot3<Scalar>>` (or something analogous for the other types and ops).

The old way of using a subclass is kept for the scalars `float` and `double` because their template specializations are identical and (unlike `Rot3`) are full (rather than partial) specializations. So there isn't an obvious way of keeping the implementation generic to avoid reuse between the two scalar types other than what was already done.