[SymForce] Separate out `create_with_derivatives`
create_with_derivatives is now two functions:
- create_with_derivatives (which computes separate jacobians for
  individual arguments, i.e. what DerivativeMode.SEPARATE_JACOBIANS did before)
- create_with_linearization (which computes either a combined jacobian,
  of the Gauss-Newton linearization)

This simplifies basically all the call sites of this function, which
typically want one of these with their respective defaults.  It also
means there's less of a combinatorial explosion of options, and the
control flow through each function is much simpler IMO.

The generated code is unchanged; I would like to also deduce the
generated function names to match what we're typically passing in, so
that we can get rid of that argument at most callsites as well.
Planning to put that in another PR.

There is a bit of duplication of code, but I think this is ok relative to the other
improvements.  This structure is also going to make it easier to
generate jacobians for functions with multiple outputs, and only support
this for `create_with_derivatives` and not `create_with_linearization`,
which is why I originally started on this.

[Testing] Existing tests

Topic: symforce-create-with-linearization
Reviewers: hayk, bradley