[SymForce] Codegen.name is snake_case, better factor name deduction
We now keep the Codegen function name as snake_case, and convert to
CamelCase when appropriate for generating C++ functions.  Also rewrote
the fancy name deduction for functions with derivatives to match how
I've been using it, which means I deleted a whole bunch of arguments
everywhere that now just use the default.

Liking the ratio of lines deleted on this one :)

[Testing] Existing tests

Topic: symforce-names
Relative: symforce-create-with-linearization