[SymForce] Nicer docstrings again
Add the outputs to the generated docstrings.  This also means that our
generated docstrings are consistent with the assumptions we make when
modifying docstrings in create_with_derivatives and
create_with_linearization (which is even more important in an upcoming PR for
derivatives of functions with multiple outputs).

[Testing] Looked at results

Topic: symforce-docstrings
Relative: symforce-names
Reviewers: hayk, bradley