[SymForce] create_with_derivatives supports multiple outputs
Motivated by camera functions which return an output and a valid flag.
This adds support for functions with multiple outputs to
`create_with_derivatives` (notably not to `create_with_linearization`,
where this definitely doesn't make sense for FULL_LINEARIZATION, and
arguably might make sense for STACKED_JACOBIAN, but would make all of
this more complicated, and I'm dubious about the use of STACKED_JACOBIAN
for use cases other than factors where the dense matmul is cheaper).

The API now allows the user to specify which outputs to compute
jacobians for, defaulting to the first output.  The jacobians are then
generated for the cartesian product of the selected given inputs and
outputs.  For functions with only one output, the semantics and defaults
should be the same as they were before.  I'll leave the rest of the
description to the docstring in the code itself.

[Testing] Added unit test

Topic: symforce-multi-jacobians
Relative: symforce-docstrings
Reviewers: hayk, bradley, himil