[SymForce] TestCaseMixin manages temporary directories
We had a bunch of tests that were either not cleaning up temporary
directories at all, or didn't use try..finally to do so.  This PR adds a
system in the SymforceTestCaseMixin for handling this automatically, so
that (1) new uses are correct by default, assuming they use the new
make_output_dir helper, (2) all existing uses are fixed, and (3) there's
less boilerplate overall.

[Testing] Verified that everything works, both with symforce TestCase
and ACBaseTest

Topic: symforce-cleanup
Relative: symforce-multi-jacobians
Reviewers: hayk, bradley