[SymForce] Clang!!
We can now build SymForce with clang!

I'm not adding a unit test to confirm that it stays this way, but I did
leave in the variable I added to Makefile.jinja which can easily be
changed to point to clang and test for the geo and cam examples.
I also have a modified version of Brian's clang compile script to filter
SymForce things and add c++14, which I can run easily.

Notably clang is more strict about having all the definitions of
things before an explicit template instantiation.

Also changing the example build to C++14 - we have some constexpr things
that C++11 was complaining about, and with C++17 or C++1z, catch2
expects to find std::conjunction but doesn't (not sure why; Godbolt
seems to think that clang-6 has it, but it's possible it's using a
different version of the standard library?).  Seems reasonable to me to
stay on C++14 anyway, and it seems like it'd be some work to make things
C++11 compatible.

clang++-6.0 does have what seems like a bug with lambda capture of
constexpr variables, which shows up in factor.tcc:176.  If I change that
line to this:
```cpp
    #pragma clang diagnostic push
    #pragma clang diagnostic ignored "-Wunused-lambda-capture"
      [func, keys_to_func, M, N](const Values<Scalar>& values, VectorX<Scalar>* residual,
    #pragma clang diagnostic pop
```
then the build passes on clang++-6.0; Godbolt indicates that this is
fixed on clang++-10.0,
so I'm not fixing this for clang++-6.0 (i.e. this PR does not in fact
make SymForce compile on clang++-6.0 without the above pragma, but I
expect it to compile on clang++-10.0 as-is).
https://godbolt.org/z/8ThWcW1KT

[Testing] Set CXX = clang++-6.0 in example/Makefile.jinja, it now
builds.  Also ran Brian's clang compile script from compile_commands.json
with clang++-6.0 -std=c++14 (on just things in SymForce), now passes

Topic: symforce-clang
Reviewers: hayk, bradley