[SymForce] Scalar handled correctly for pow
The interface to std::pow does not actually allow for templating on the
desired type, and instead promotes the arguments to `double` if either
is integral.  So, if we want to do everything with type Scalar, we need
to make sure to cast both arguments to Scalar first.

(Also we had lingering double literals in Pow(expr, -1) and Heaviside(expr))

[Testing] Regenerated code and inspected; some playing around on Godbolt

Topic: symforce-moar-scalar
Reviewers: hayk, bradley
Relative: symforce-multi-jacobians