[SymForce] Improve Rot2.Inverse
Previously we calculated the inverse of a `Rot2` by taking the inverse of the complex number it is represented by. But `Rot2`s are represented by unit complex numbers, so inverting the magnitude of the complex number is unnecessary (as we are just multiplying by 1), so we just need to take the complex conjugate.

The only manual change made in this PR is changing (in `third_party_modules/symforce/symforce/geo/rot2.py`)
``` python
def inverse(self) -> Rot2:
    return self.__class__(self.z.inverse())
```
to
``` python
def inverse(self) -> Rot2:
    return self.__class__(self.z.conj())
```