[SymForce] Cholesky ordering is a runtime parameter
The ordering for the SparseCholeskySolver is now a runtime parameter (a
functor stored as a std::function) instead of a template parameter.
This means that the vast majority of applications can now store their
optimizer as a sym::Optimizerd or sym::Optimizerf, and use the GNC
subclass if desired, or a different variable ordering, but will rarely
need to need to change template parameters or be unable to switch things
around based on runtime parameters.

[Testing] Existing tests, added test that sym::Optimizer constructor
compiles (since it now nontrivially forwards the nonlinear solver
constructor args)

Topic: symforce-runtime-ordering
Relative: symforce-optimizer-constructor-name
Reviewers: hayk, bradley, ryan