[SymForce] sm.solve API is the same(ish) on SymPy and SymEngine
Annoyingly, the SymPy and SymEngine versions of sm.solve return
different types for a finite set of solutions (SymPy returns a list,
SymEngine returns a FiniteSet).  I think we don't want to modify the
SymPy API since it's likely more powerful and we don't want to restrict
expressiveness there, so this adds a translation of the SymEngine
results to a tuple.  Won't work for infinite solution sets, but works
for empty and finite sets

[Testing] Added unit test

Reviewers: hayk, bradley
Topic: symforce-solve