[SymForce] codegen Compose and FromYawPitchRoll
This PR changes `Rot3::FromYawPitchRoll` and the `Compose` methods for `Rot2`, `Rot3`, `Pose2`, and `Pose3` as requested.

While I liked the idea of reducing the number of different implementations we had floating around, I wanted to make sure I wasn't doing anything to dramatically hurt performance, so I looked a bit into it.

When benchmarked* (compiled with `--ffast-math` and `-march=native`):
`Pose3<float>::Compose` was ~7 ns when handwritten, and ~4 ns when generated
`Pose3<double>::Compose` was ~13 ns when handwritten, and ~3.5 ns when generated

`Rot3<float>::FromYawPitchRoll` was ~25 ns when handwritten, and ~26 ns when generated
`Rot3<double>:FromYawPitchRoll` was ~61 ns when handwritten, and ~63 ns when generated.

*with the disclaimer that I'm a novice when it comes to proper micro-benchmarking