[SymForce] Update tutorials
Update the tutorials to align with current recommended practice of
generating from Python functions instead of input/output Values.  Also
fixes the matplotlib warnings in the sympy tutorial. The
optimization tutorial is still out of date, but now points to the bundle
adjustment example, with a TODO to either use cling or update with a
python optimizer.

Reviewers: hayk, bradley
Topic: symforce-tutorials