[SymForce] Calculate Jacobians using first order approximations instead of chain rule
The non-generated change was to `tangent_jacobians` in `jacobian_helpers.py`.

This form of `tangent_jacobians` is the same as what we wrote after the Jedi council, except rather than using infinitesimal exp and log, it uses infinitesimal retract and local_coordinates. This is because `a.retract(t)` does not always equal `a * exp(t)` (see `Pose2` and `Pose3`) (and something analogous for `local_coordinates`). Since we want our Jacobians to be consistent with these functions, we need to be careful to not make assumptions about the form of `retract` and `local_coordinates`.

The overall idea is still the same, however. We used the chain rule because it is `retract` and `local_coordiantes` are often too complicated to compose with our function and then differentiate. However, that has the downside of distributing (in the sense of `a(x + y) = ax + ay`) many of the operations. This was bad for cse, bloating the number of operations we needed.

The alternative to the chain rule is to replace `retract` and `local_coordinates` with simple first order approximations which we can easily differentiate. This way we can compose the functions and still take the derivative. Since no matrix multiplication was involved, our expressions remain factored, and cse is able to simplify the expressions more easily.