Make Github review hide changes in generated code Topic: gitattributes
As explained [here](https://docs.github.com/en/github/administering-a-repository/managing-repository-settings/customizing-how-changed-files-appear-on-github), GitHub uses .gitattributes files to know whether code files are auto-generated, so it can show them collapsed by default in reviews and exclude them from repository language statistics (i.e. it's not useful to know our repo is 50% JSON). This review adds a number of `.gitattributes` files to document code that should be considered generated for the purposes of code reviews.

See the below relative review for an example of what this looks like.

Reviewers: nikhil,josiah,abe,hayk