[SymForce] Add sym::LieGroupOps<T>::IsClose
**Introduction**
The problem this is meant to address is that we've been using `IsApprox` in a variety of places to test if two group elements are nearly the same. This is unsound, however, as `sym::Rot3f(Eigen::Quaternion<float>(1, 0, 0, 0))` and `sym::Rot3f(Eigen::Quaternion<float>(-1, 0, 0, 0))` represent the exact same rotation despite `IsApprox` returning `false`.

`IsClose` instead checks that the two group elements are close by asking for the log of their quotient, then checking that the log is nearly the zero vector. While, more or less, the log is a good measure of distance, we need to take care with the multi-valued nature of the log.

Two elements are close if their quotient is close to the identity. An element g is close to the identity if there exists t in the tangent space of the identity such that exp(t) = g and |t| is small. Hence, while log(g) might have multiple values, we only care about the shortest one, as the longer ones just mean there is another path parameterized by the exponential map that passes through g that is longer than the shortest one.

This is all to say that if a g is "far" from the identity, it doesn't matter which value of log(g) we return, as none of them are small. So, regardless of the implementation of `ToTangent` `IsClose` will correctly conclude that two elements are far.

The implementation is, however, sensitive to which log is returned if two elements are close. While the smallest log(1 + i\*epsilon) is epsilon, we can find another log of size 2*pi + epsilon. One is small, and the other is not.

So, it's important that `ToTangent` return the smallest (or, a tangent not much larger than the smallest) tangent for group elements near the identity.

**Notes on Implementation**
I took Aaron's advice and implemented `IsClose` in C++ and not Python.

Since this one implementation should work for all lie groups, I put the generic implementation in a helper class `LieGroupOpsImpls` so that the method implementation can be added to any specialization of `LieGroupOps<T>` without having to repeat myself too much. This also has the benefit that if we had another generic implementation we wanted to add to `LieGroupOps`, we can just implement it in `LieGroupOpsImpls` and it will be automatically added to all the specializations. Moreover, if a specialization wants some default implementations but not others, it can just redefine the default implementation it wants to specialize (even deleting it for the specialization if we wanted to).

I put `LieGroupOpsImpls` in the namespace `default_base` to make it more clear that if one wants the function `IsClose` one should use `LieGroupOps` and not `default_base::LieGroupOpsImpls`. I could have made `IsClose` a protected method and then call it within the specialization to prevent anyone from being able to use `LieGroupOpsImpls` directly, but that would have required a lot more boilerplate.

**Testing**
This I was less clear on how to best proceed. I wanted the tests of `IsClose` to be with the other `LieGroupOps` tests, but the problem was that while the implementation of `IsClose` is easy enough to write generically, the tests one would want to have depend very much of the group itself and its implementation. For example, testing that `IsClose(Rot3(q), and Rot3(-q))` (this is pseudo code) returns `true` if `q` is a quaternion.

To handle this, I wanted to have a templated function which would have a generic test that would make sense for all Lie groups, but which could be specialized for `Rot3` and `Rot2` (and any other group we'd want). The problem is that C++ doesn't allow partial function template specializations, and I wanted to make the test for `Rot3` and `Rot2` generic in the scalar type used. To address this, I instead put the test in a struct (which can be partially specialized).

I also modified some tests at the top of `geo_package_cpp_test.cc` to use `IsClose` instead of `IsApprox` because they were causing me issues when they erroneously failed after I made a modification to the way random rot3's are generated (they falsely said that `Rot3(-q)` and `Rot(q)` are not the same rotation).